Support pytest (#61)
* Support pytest

* Cosmetic changes

* undo merge mistakes

* Revert "Merge branch 'master' into EJBfixes"

This reverts commit b369bd55ec78a81644d0fadf7cb006b503bf6a09, reversing
changes made to 79ab56f4c2435dc0a91dca8760f02d69a2d2a495.

+++ pyspike/cython/cython_simulated_annealing.pyx
@@ -38,7 +38,7 @@ from libc.math cimport fmod
 from libc.stdlib cimport rand
 from libc.stdlib cimport RAND_MAX
 
-DTYPE = np.float
+DTYPE = float
 ctypedef np.float_t DTYPE_t
